Fix an error, Added the cast from int32 to byte in the return.

+++ DomainTests/RecordGenerator.cs
@@ -471,7 +471,7 @@ internal class RecordGenerator
                 }
                 else
                 {
-                    return 0;
+                    return (byte)0;
                 }
             })
             .RuleFor(x => x.GreenHouse, (f, u) => _greenHouses.FirstOrDefault(x => x.Id == u.GreenHouseId, null))
